+++ gvmtools/helper.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2018 Greenbone Networks GmbH
+# Copyright (C) 2018-2021 Greenbone Networks GmbH
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 #
@@ -19,7 +19,10 @@
 import getpass
 import os
 import sys
+import uuid
+import string
 
+from random import choice, randrange
 from lxml import etree
 
 from gvm.errors import GvmError
@@ -111,6 +114,33 @@ def error_and_exit(msg):
     """
     print("\nError: {}\n".format(msg), file=sys.stderr)
     sys.exit(1)
+
+
+def generate_random_ips(count: int):
+    """Generate count random IPv4s"""
+    exclude_127 = [i for i in range(1, 256)]
+    exclude_127.remove(127)
+    return [
+        '{}.{}.{}.{}'.format(
+            choice(exclude_127),
+            randrange(0, 256),
+            randrange(0, 256),
+            randrange(1, 255),
+        )
+        for i in range(count)
+    ]
+
+
+def generate_id(
+    size: int = 12, chars: str = string.ascii_uppercase + string.digits
+):
+    """Generate a random ID"""
+    return ''.join(choice(chars) for _ in range(size))
+
+
+def generate_uuid():
+    """Generate a random new uuid"""
+    return str(uuid.uuid4())
 
 
 def create_xml_tree(xml_doc):