+++ scripts/create-dummy-data.gmp.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2017-2019 Greenbone Networks GmbH
+# Copyright (C) 2017-2021 Greenbone Networks GmbH
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 #
@@ -16,8 +16,11 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import random
-import string
+import sys
+
+from random import choice
+
+from gvmtools.helper import generate_id
 
 
 def check_args(args):
@@ -33,11 +36,7 @@ def check_args(args):
     ssh --hostname <gsm> scripts/create-dummy-data.gmp.py <count>
         """
         print(message)
-        quit()
-
-
-def id_generator(size=12, chars=string.ascii_uppercase + string.digits):
-    return ''.join(random.choice(chars) for _ in range(size))
+        sys.exit()
 
 
 def create_data(gmp, count):
@@ -45,7 +44,7 @@ def create_data(gmp, count):
     target_ids = []
 
     for _ in range(0, count):
-        name = id_generator()
+        name = generate_id()
         gmp.create_credential(
             name,
             login=name,
@@ -55,27 +54,27 @@ def create_data(gmp, count):
     print(str(count) + ' random credentials generated.')
 
     for _ in range(0, count):
-        name = id_generator()
+        name = generate_id()
         gmp.create_port_list(name, port_range='T:1-42')
     print(str(count) + ' random port lists generated.')
 
     for _ in range(0, count):
-        name = id_generator()
+        name = generate_id()
         res = gmp.create_config('085569ce-73ed-11df-83c3-002264764cea', name)
         config_ids.append(res.xpath('@id')[0])
     print(str(count) + ' random scan configs generated.')
 
     for _ in range(0, count):
-        name = id_generator()
+        name = generate_id()
         res = gmp.create_target(name, hosts=['127.0.0.1'])
 
         target_ids.append(res.xpath('@id')[0])
     print(str(count) + ' random targets generated.')
 
     for _ in range(0, count):
-        name = id_generator()
-        config_id = random.choice(config_ids)
-        target_id = random.choice(target_ids)
+        name = generate_id()
+        config_id = choice(config_ids)
+        target_id = choice(target_ids)
         gmp.create_task(
             name, config_id, target_id, '08b69003-5fc2-4037-a479-93b440211c73'
         )