+++ scripts/pdf-report.gmp.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2019 Greenbone Networks GmbH
+# Copyright (C) 2019-2021 Greenbone Networks GmbH
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 #
@@ -38,7 +38,7 @@ def check_args(args):
 ssh --hostname <gsm> scripts/pdf-report.gmp.py <report_id> <pdf_file>
         """
         print(message)
-        quit()
+        sys.exit()
 
 
 def main(gmp, args):