+++ scripts/start-nvt-scan.gmp.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2017-2019 Greenbone Networks GmbH
+# Copyright (C) 2017-2021 Greenbone Networks GmbH
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 #
@@ -15,6 +15,8 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+import sys
 
 
 def check_args(args):
@@ -32,7 +34,7 @@ ssh --hostname <gsm> scripts/start-nvt-scan.gmp.py \
     1.3.6.1.4.1.25623.1.0.106223 localhost
         """
         print(message)
-        quit()
+        sys.exit()
 
 
 def get_config(gmp, nvt_oid):